+++ src/Project_3d/Frustum.cpp
@@ -262,7 +262,7 @@ bool mydx::Frustum::Initialize(Graphics& graphics) noexcept
 
     AddBind(std::make_unique<VertexBuffer>(graphics, vertices));
 
-    const std::vector<unsigned short> indices =
+    const std::vector<DWORD> indices =
     {
         0,2,1, 2,3,1,
         1,3,5, 3,7,5,
